backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.8
requests==2.31.0
python-multipart==0.0.6
cors==1.0.1
